refactor(#8/comment): 댓글 관련 exception을 application -> domain으로 변경
